requirements-dev.txt
Django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-filter==2.4.0
djoser==2.1.0
python-decouple==3.4
dj-database-url==0.5.0
dj-static==0.0.6
django-rest-resetpassword==0.1.2
drf-yasg==1.20.0
mysqlclient==2.0.3
